Strip AbstractMasterPopTableFactory
It only ever had two concrete implementations, and one of those is now
long gone. Hard-code to use the concrete everywhere.
